Refactor: cts: Move remote.py into the _cts module.
