+++ cts/lab/ClusterManager.py
@@ -17,12 +17,12 @@ from cts.CIB         import ConfigFactory
 from cts.CTS         import NodeStatus, Process
 from cts.CTStests    import AuditResource
 from cts.watcher     import LogWatcher
-from cts.remote      import RemoteFactory
 from cts.environment import EnvFactory
 
 from pacemaker.buildoptions import BuildOptions
 from pacemaker._cts.logging import LogFactory
 from pacemaker._cts.patterns import PatternSelector
+from pacemaker._cts.remote import RemoteFactory
 
 has_log_stats = {}
 log_stats_bin = BuildOptions.DAEMON_DIR + "/cts_log_stats.sh"
