+++ cts/lab/__init__.py
@@ -12,6 +12,5 @@ CTSscenarios
 CTStests
 environment
 patterns
-remote
 watcher
 """
